Fix #17 Panic in RtcServer
Before attempting to create the RTC listener, check
if we can bind to UDP (not TCP - this is the bug) to
find an available port in the range.

Adds CLI parsing to the demo server. Exercise the
change (from demo/server) with:

`cargo run --features use-webrtc -- 14191`
`cargo run --features use-webrtc -- 14192`

Observe that the first RTC server starts on 8000,
while the second will now start on 8001, instead
of crashing.